refractor: dup. code reduced to util function
