+++ frontend/src/components/RoomCard.tsx
@@ -1,12 +1,12 @@
 import "./styles/RoomCard.css";
 import { RoomModel } from "./model/RoomModel.ts";
 import { useNavigate } from "react-router-dom";
-import axios from "axios";
+import handleToggleWishlist from "../utils/handleToggleWishlist.ts";
 
 type RoomCardProps = {
     room: RoomModel;
     user: string;
-    onStatusChange?: (updatedRoom: RoomModel) => void;
+    onStatusChange: (updatedRoom: RoomModel) => void;
 };
 
 export default function RoomCard(props: Readonly<RoomCardProps>) {
@@ -16,21 +16,6 @@ export default function RoomCard(props: Readonly<RoomCardProps>) {
         navigate(`/room/${props.room.id}`);
     };
 
-    const handleToggleWishlist = () => {
-        const updatedStatus =
-            props.room.wishlistStatus === "ON_WISHLIST" ? "NOT_ON_WISHLIST" : "ON_WISHLIST";
-
-        const updatedRoom = { ...props.room, wishlistStatus: updatedStatus };
-
-        axios
-            .put(`/api/practice-hub/${props.room.id}`, updatedRoom)
-            .then((response) => {
-                if (props.onStatusChange) {
-                    props.onStatusChange(response.data);
-                }
-            })
-            .catch((error) => console.error("Error updating wishlist status:", error));
-    };
 
     return (
         <div className="room-card" onClick={handleCardClick} style={{ cursor: "pointer" }}>
@@ -44,7 +29,7 @@ export default function RoomCard(props: Readonly<RoomCardProps>) {
                     id="button-wishlist"
                     onClick={(event) => {
                         event.stopPropagation(); // Verhindert die Weitergabe des Klicks an die Karte
-                        handleToggleWishlist();
+                        handleToggleWishlist(props.room, props.onStatusChange);
                     }}
                     className={props.room.wishlistStatus === "ON_WISHLIST" ? "wishlist-on" : "wishlist-off"}
                 >
